refactoring the filter operations for the map controller
