+++ src/main/java/eu/ec/doris/kohesio/controller/ProjectController.java
@@ -5,7 +5,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import eu.ec.doris.kohesio.payload.*;
+import eu.ec.doris.kohesio.services.FiltersGenerator;
 import eu.ec.doris.kohesio.services.SPARQLQueryService;
+import eu.ec.doris.kohesio.services.SimilarityService;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
 import org.apache.commons.math3.util.Precision;
@@ -47,6 +49,12 @@ public class ProjectController {
 
     @Autowired
     SPARQLQueryService sparqlQueryService;
+
+    @Autowired
+    SimilarityService similarityService;
+
+    @Autowired
+    FiltersGenerator filtersGenerator;
 
     @Value("${kohesio.sparqlEndpoint}")
     String sparqlEndpoint;
@@ -514,8 +522,11 @@ public class ProjectController {
         if (offset < 1000) {
             offset = 0;
             limit = 1000;
+            if(keywords != null){
+                limit = 100;
+            }
         }
-        String search = filterProject(keywords, country, theme, fund, program, categoryOfIntervention, policyObjective, budgetBiggerThen, budgetSmallerThen, budgetEUBiggerThen, budgetEUSmallerThen, startDateBefore, startDateAfter, endDateBefore, endDateAfter, latitude, longitude, region, limit, offset);
+        String search = filtersGenerator.filterProject(keywords, country, theme, fund, program, categoryOfIntervention, policyObjective, budgetBiggerThen, budgetSmallerThen, budgetEUBiggerThen, budgetEUSmallerThen, startDateBefore, startDateAfter, endDateBefore, endDateAfter, latitude, longitude, region,null,limit, offset);
 
         int numResults = 0;
 
@@ -720,7 +731,10 @@ public class ProjectController {
             resultList.add(project);
         }
         ProjectList projectList = new ProjectList();
-        if(offset <= 990) {
+        int upperLimit = 990;
+        if(keywords != null)
+            upperLimit = 90;
+        if(offset <= upperLimit) {
             for (int i = inputOffset; i < Math.min(resultList.size(), inputOffset + inputLimit); i++) {
                 projectList.getList().add(resultList.get(i));
             }
@@ -730,33 +744,7 @@ public class ProjectController {
         projectList.setNumberResults(numResults);
         return new ResponseEntity<ProjectList>(projectList, HttpStatus.OK);
     }
-    /*
-    Query the similarity API and get a list of similar words
-     */
-    private ArrayList<SimilarWord> getSimilarWords(String text){
-        String url = "http://similarity.cnect.eu:3000/similarity?text="+text;
-        ArrayList<SimilarWord> similarWords = new ArrayList<>();
-        try {
-            RestTemplate restTemplate = new RestTemplate();
-            ResponseEntity<String> response = restTemplate.getForEntity(url, String.class);
-            if (response.getStatusCode().equals(HttpStatus.OK)) {
-                ObjectMapper mapper = new ObjectMapper();
-                JsonNode root = mapper.readTree(response.getBody());
-                ObjectNode result = (ObjectNode) root;
-                ArrayNode hits = (ArrayNode) result.get("similar");
-                for (int i = 0; i < hits.size(); i++) {
-                    String word = hits.get(i).get("word").textValue();
-                    double score  = hits.get(i).get("score").doubleValue();
-                    similarWords.add(new SimilarWord(word.replace("_"," "),score));
-                }
-            } else {
-                System.err.println("Error in HTTP request!");
-            }
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return similarWords;
-    }
+
     private SemanticSearchResult getProjectsURIsfromSemanticSearch(String keywords,boolean cache,int offset,int limit) {
         String url = null;
         //String encoded = URLEncoder.encode(keywords, StandardCharsets.UTF_8.toString());
@@ -842,7 +830,7 @@ public class ProjectController {
             Principal principal)
             throws Exception {
         logger.info("language {} keywords {} country {} theme {} fund {} program {} categoryOfIntervention {} policyObjective {} budgetBiggerThen {} budgetSmallerThen {} budgetEUBiggerThen {} budgetEUSmallerThen {} startDateBefore {} startDateAfter {} endDateBefore {} endDateAfter {} latitude {} longitude {} region {} limit {} offset {} granularityRegion {}", language, keywords, country, theme, fund, program, categoryOfIntervention, policyObjective, budgetBiggerThen, budgetSmallerThen, budgetEUBiggerThen, budgetEUSmallerThen, startDateBefore, startDateAfter, endDateBefore, endDateAfter, latitude, longitude, region, limit, offset, null);
-        String search = filterProject(keywords, country, theme, fund, program, categoryOfIntervention, policyObjective, budgetBiggerThen, budgetSmallerThen, budgetEUBiggerThen, budgetEUSmallerThen, startDateBefore, startDateAfter, endDateBefore, endDateAfter, latitude, longitude, region, limit, offset);
+        String search = filtersGenerator.filterProject(keywords, country, theme, fund, program, categoryOfIntervention, policyObjective, budgetBiggerThen, budgetSmallerThen, budgetEUBiggerThen, budgetEUSmallerThen, startDateBefore, startDateAfter, endDateBefore, endDateAfter, latitude, longitude, region,null, limit, offset);
 
         //computing the number of results
         String searchCount = search;
@@ -1047,177 +1035,5 @@ public class ProjectController {
         }
     }
 
-    private String filterProject(String keywords, String country, String theme, String fund, String program, String categoryOfIntervention,
-                                 String policyObjective, Integer budgetBiggerThen, Integer budgetSmallerThen, Integer budgetEUBiggerThen, Integer budgetEUSmallerThen, String startDateBefore, String startDateAfter,
-                                 String endDateBefore,
-                                 String endDateAfter,
-                                 String latitude,
-                                 String longitude,
-                                 String region,
-                                 Integer limit,
-                                 Integer offset) throws IOException {
-        String search = "";
-        if (keywords != null) {
-//            if (!keywords.contains("AND") && !keywords.contains("OR") && !keywords.contains("NOT")) {
-//                String[] words = keywords.split(" ");
-//                StringBuilder keywordsBuilder = new StringBuilder();
-//                for (int i = 0; i < words.length - 1; i++) {
-//                    keywordsBuilder.append(words[i]).append(" AND ");
-//                }
-//                keywordsBuilder.append(words[words.length - 1]);
-//                keywords = keywordsBuilder.toString();
-//            }
-            String[] words = keywords.split(" ");
-            StringBuilder keywordsBuilder = new StringBuilder();
-            keywordsBuilder.append("(");
-            for (int i = 0; i < words.length; i++) {
-                if(i < words.length -1){
-                    keywordsBuilder.append(words[i]).append(" AND ");
-                }else{
-                    keywordsBuilder.append(words[i]);
-                }
-            }
 
-            keywordsBuilder.append(")").append(" OR \"").append(keywords).append("\"^2");
-
-            ArrayList<SimilarWord> similarWords = getSimilarWords(keywords);
-            for (SimilarWord word: similarWords) {
-                keywordsBuilder.append(" OR \"").append(word.getWord()).append("\"^").append(word.getScore());
-            }
-            String newQuery = keywordsBuilder.toString();
-
-            search +=
-                    "?s0 <http://www.openrdf.org/contrib/lucenesail#matches> [ "
-                            + "<http://www.openrdf.org/contrib/lucenesail#query> \""
-                            + newQuery.replace("\"", "\\\"")
-                            + "\" ] .";
-
-        }
-//        if(keywords != null) {
-//            SemanticSearchResult semanticSearchResult = getProjectsURIsfromSemanticSearch(keywords, false, 0, 5000);
-//            ArrayList<String> projectsURIs = semanticSearchResult.getProjectsURIs();
-//            if (projectsURIs.size() > 0) {
-//                //search = "";
-//                search += "VALUES ?s0 {";
-//                for (String uri : projectsURIs) {
-//                    String uriStr = "<" + uri + ">";
-//                    search += uriStr + " ";
-//                }
-//                search += "}";
-//                //numResults = semanticSearchResult.getNumberOfResults();
-//            } else {
-//                System.out.println("Semantic search API returned empty result!!");
-//            }
-//        }
-        if (country != null && region == null) {
-            search += "?s0 <https://linkedopendata.eu/prop/direct/P32> <" + country + "> . ";
-        }
-
-        if (theme != null) {
-            search +=
-                    "?s0 <https://linkedopendata.eu/prop/direct/P888> ?category. "
-                            + "?category <https://linkedopendata.eu/prop/direct/P1848> <"
-                            + theme
-                            + "> . ";
-        }
-
-        if (policyObjective != null) {
-            search +=
-                    "?s0 <https://linkedopendata.eu/prop/direct/P888> ?category. "
-                            + "?category <https://linkedopendata.eu/prop/direct/P1849> <"
-                            + policyObjective
-                            + "> . ";
-        }
-
-        if (fund != null) {
-            search += "?s0 <https://linkedopendata.eu/prop/direct/P1584> <" + fund + "> . ";
-        }
-
-        if (program != null) {
-            search += "?s0 <https://linkedopendata.eu/prop/direct/P1368> <" + program + "> . ";
-        }
-
-        if (categoryOfIntervention != null) {
-            search +=
-                    "?s0 <https://linkedopendata.eu/prop/direct/P888> <" + categoryOfIntervention + "> . ";
-        }
-
-        if (budgetBiggerThen != null) {
-            search +=
-                    " ?s0 <https://linkedopendata.eu/prop/direct/P474> ?budget . "
-                            + "FILTER( ?budget > "
-                            + budgetBiggerThen
-                            + ")";
-        }
-
-        if (budgetSmallerThen != null || budgetBiggerThen != null) {
-            search += " ?s0 <https://linkedopendata.eu/prop/direct/P474> ?budget . ";
-            if (budgetBiggerThen != null) {
-                search += "FILTER( ?budget > " + budgetBiggerThen + ")";
-            }
-            if (budgetSmallerThen != null) {
-                search += "FILTER( ?budget < " + budgetSmallerThen + ")";
-            }
-        }
-
-        if (budgetEUBiggerThen != null || budgetEUSmallerThen != null) {
-            search += " ?s0 <https://linkedopendata.eu/prop/direct/P835> ?budgetEU . ";
-            if (budgetEUBiggerThen != null) {
-                search += "FILTER( ?budgetEU > " + budgetEUBiggerThen + ")";
-            }
-            if (budgetEUSmallerThen != null) {
-                search += "FILTER( ?budgetEU < " + budgetEUSmallerThen + ")";
-            }
-        }
-
-        if (startDateBefore != null || startDateAfter != null) {
-            search += " ?s0 <https://linkedopendata.eu/prop/direct/P20> ?startDate . ";
-            if (startDateBefore != null) {
-                search +=
-                        "FILTER( ?startDate <= \""
-                                + startDateBefore
-                                + "T00:00:00Z\"^^<http://www.w3.org/2001/XMLSchema#dateTime>)";
-            }
-            if (startDateAfter != null) {
-                search +=
-                        "FILTER( ?startDate >= \""
-                                + startDateAfter
-                                + "T00:00:00Z\"^^<http://www.w3.org/2001/XMLSchema#dateTime>)";
-            }
-        }
-
-        if (endDateBefore != null || endDateAfter != null) {
-            search += " ?s0 <https://linkedopendata.eu/prop/direct/P33> ?endDate . ";
-            if (endDateBefore != null) {
-                search +=
-                        "FILTER( ?endDate <= \""
-                                + endDateBefore
-                                + "T00:00:00Z\"^^<http://www.w3.org/2001/XMLSchema#dateTime>)";
-            }
-            if (endDateAfter != null) {
-                search +=
-                        "FILTER( ?endDate >= \""
-                                + endDateAfter
-                                + "T00:00:00Z\"^^<http://www.w3.org/2001/XMLSchema#dateTime>)";
-            }
-        }
-
-        if (region != null) {
-            search += "?s0 <https://linkedopendata.eu/prop/direct/P1845> <" + region + "> . ";
-        }
-
-        if (latitude != null && longitude != null) {
-            search +=
-                    "?s0 <https://linkedopendata.eu/prop/direct/P127> ?coordinates . "
-                            + "FILTER ( <http://www.opengis.net/def/function/geosparql/distance>(\"POINT("
-                            + longitude
-                            + " "
-                            + latitude
-                            + ")\"^^<http://www.opengis.net/ont/geosparql#wktLiteral>,?coordinates,<http://www.opengis.net/def/uom/OGC/1.0/metre>)< 100000) . ";
-        }
-
-        search +=
-                "   ?s0 <https://linkedopendata.eu/prop/direct/P35> <https://linkedopendata.eu/entity/Q9934> . ";
-        return search;
-    }
 }
